+++ Gateway/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+
+  AES_IV  : 'YWFhYWFhYWFhYWFhYWFhYssQ==',
+  //'AAAAAAAAAAAAAAAAAAAAAA==';//'YWFhYWFhYWFhYWFhYWFhYssQ=='
+  AES_KEY : '6162636465666768696a6b6c6d6e6f707172737475767778797a313233343536',
+  //'2B7E151628AED2A6ABF7158809CF4F3C2B7E151628AED2A6ABF7158809CF4F3C';
+
+  SER_PORT : '/dev/cu.usbmodem1411'
+
+}
